backend/requirements.txt
anthropic==0.54.0
python-dotenv==1.0.0
pillow==10.1.0
pyautogui==0.9.54
flask==3.0.0
flask-cors==4.0.0
httpx<0.28
waitress==3.0.1 